+++ NEWS.md
@@ -1,3 +1,9 @@
+## detectPVC 0.1.17 (2024-09-28)
+
+- Added function plot_states() for classifying PVC state as
+  normal/bigeminy/trigeminy/omitted and making a plot
+
+
 ## detectPVC 0.1.16 (2024-05-26)
 
 - A new package, to detect premature ventricular complexes (PVCs) in